requirements.txt
scapy
pysnmp
python-nmap